+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `splitLines` string utility function
 - `tryParseSwissIbanNumber` swiss standard function
+- `tryParseSwissSocialInsuranceNumber` swiss standard function
+- `splitLines` string utility function
 - `trimStart`, `trimEnd` and `trim` string type utility functions
 
 ### Changed